SAMPLE: Removed tests and modified scanner.

+++ SAMPLE/Main.java
@@ -6,7 +6,7 @@ public class Main {
 	public static void main(String args[]) throws Exception {		
 		Scanner scanner = new Scanner(System.in);
 
-		long testCases = scanner.nextLong();
+		int testCases = scanner.nextInt();
 		for(int i = 0; i < testCases; i++) {
 			System.out.println(i + 1);
 		}
